move job determination functions to separate class
